+++ src/types.ts
@@ -9,10 +9,6 @@ export interface HydrationParams {
   [key: string]: string;
 }
 
-export interface PlaylistSchedule {
-  byDisplayAt?: boolean;
-}
-
 export interface PlaylistItem {
   source: string;
   displayAt?: string;
@@ -23,6 +19,5 @@ export interface Playlist {
   dpVersion: string;
   title: string;
   items: PlaylistItem[];
-  schedule?: PlaylistSchedule;
   [key: string]: unknown;
 }